crates/llm/pyproject.toml
[tool.poetry]
name = "llm_handler"
version = "0.1.0"
description = "用於 Rust 的 Python 模組"
authors = ["liminchian <liminchian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-cpp-python = {version = "^0.2.82", source = "llama-cpp-python"}

[[tool.poetry.source]]
name = "llama-cpp-python"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

